+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -9,6 +9,7 @@
 
 #### Fixed
 - Inconsistent completion list with implicitly defined private CSECT
+- Incorrect results of string comparison in CA expressions
 
 ## [1.11.1](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.11.0...1.11.1) (2023-12-04)